Chore: renamed user authentication service to component
